Rename ProfileConstants to ExtraConstants and move it to extra package Add the 3rd constant
