+++ app_pojavlauncher/src/main/java/net/kdt/pojavlaunch/extra/ExtraConstants.java
@@ -1,9 +1,10 @@
-package net.kdt.pojavlaunch.profiles;
+package net.kdt.pojavlaunch.extra;
 
-public class ProfileConstants {
+public class ExtraConstants {
     /* ExtraCore constant: a HashMap for converting values such as latest-snapshot or latest-release to actual game version names */
     public static final String RELEASE_TABLE = "release_table";
     /* ExtraCore constant: an ArrayList of Strings, where each String is a Minecraft version name */
     public static final String VERSION_LIST= "lac_version_list";
-
+    /* ExraCore constant: Serpent's back button tracking thing */
+    public static final String BACK_PREFERENCE = "back_preference";
 }
